requirements.txt
pip~=23.2.1
attrs~=25.1.0
wheel~=0.41.2
protobuf~=5.29.3
cryptography~=44.0.0
pillow~=11.1.0
tornado~=6.4.2
Jinja2~=3.1.5
setuptools~=68.2.0
passlib~=1.7.4
python-jose~=3.3.0
fastapi~=0.115.8
pydantic~=2.10.6
SQLAlchemy~=2.0.38
streamlit~=1.42.0
requests~=2.32.3